1.21.3 cursed port (read desc)
Flag rendering is no longer visually the same in terms of behaviour. Dropped items all use the same renderer seed. Flag randomization is based off... damage to items.

Arms do not hold the flag as they did previously.

Item rendering is kinda hijacked for this, which makes me uncomfortable in terms of mod compat. (shaders did work when tested)

there's probably more to it, but that's the biggest changes I've made to this branch.